server/requirements.txt
fastapi
uvicorn
python-dotenv
openai
firebase_admin
python-multipart